+++ changelog.md
@@ -1,6 +1,7 @@
 # 0.6.0 (WIP)
 
 - Increase Arduino code size from 128KB to 256KB
+- Increase Arduino SRAM from 16KB to 20KB
 - rename menu Featherlib (Beta) --> Featherlib (Debug)
 - print warning message when Featherlib & Arduino library version are not matched
 - fix get TCP peer information when SOCKET is CLOSED